Version 1.52
Critical bug FIXED: Alignment of code failed on indents with tabs. Realized that stridx() and matchend() read a tab as 1 space, instead of the correct &tabstop number of spaces. Thanks to Christian Brehm!

Renamed options to g:vhdl_indent_genportmap & g:vhdl_indent_rhsassign, to be consistent with naming convention.